refactor(ui): extract tool dispatcher + display chrome to dedicated module
Slice 11b.1 step 8 of the message-part.tsx split. Pull the assistant-side
tool dispatcher (PART_MAPPING["tool"]) plus the file-modifying display
chrome it shares with the legacy tool renderers into a focused module so
message-part.tsx keeps shrinking toward the AGENTS.md 500-line ceiling.

New: packages/ui/src/components/message-part-tool-display.tsx
  - ShellSubmessage              animated subtitle used by bash + diff cards
  - Diagnostic / getDiagnostics  LSP diagnostic filter
  - DiagnosticsDisplay           rendered banner for edit / write / apply_patch
  - ToolFileAccordion            sticky-header accordion for file tools
  - side-effect                  registers PART_MAPPING["tool"]

New: packages/ui/src/components/message-part-session-link.ts
  - sessionLink                  pathname-aware child session URL builder
  - currentSession               parse parent session id from current path
  - taskSession                  resolve child session id for task/agent tool

The tool dispatcher and the still-resident task renderer both consume
these routing helpers — pulling them into a pure module avoids dragging
Kobalte / motion side-effects through tools-agent when that module lands
next.

Source-grep invariants (message-part-stale.test.ts) re-anchored to the
new file locations; behaviour assertions unchanged.

tsgo --noEmit GREEN. bun test: 478 pass / 3 pre-existing fail
(Button/IconButton fixture rot). app tests: 1050 pass / 0 fail.
